// frontend/src/types/generated/physiologicLimits.ts
export interface PhysiologicLimit {
  min: number;
  max: number;
  unit?: string;
  description?: string;
}